+++ CHANGELOG.md
@@ -17,6 +17,7 @@ message.
 - You now need to do `assert(value).all { ... }` instead of `assert(value) { ... }`. This makes the implementation more
 consistent and better mirrors `assertAll { ... }` which has a similar behavior.
 - New dependency on kotlin-reflect
+- `contains`, `doesNotContain`, and `each` works on Iterable instead of Collection
 
 ### Breaking Changes
 - Because of the above, `Assert<Collection>.all` and `Assert<Array>.all` have both been renamed to `each`.